requirements.txt
# autopep8==1.6.0
# cycler==0.11.0
# kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.5
# Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
# pytz==2021.3
scipy==1.7.1
# six==1.16.0
# toml==0.10.2
